fixed some warnings. use size_t for some array indexes. use int for CString indexes.
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4970 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/SubPic/DX9SubPic.cpp
@@ -367,8 +367,8 @@ CDX9SubPicAllocator::~CDX9SubPicAllocator()
 void CDX9SubPicAllocator::GetStats(int &_nFree, int &_nAlloc)
 {
 	CAutoLock Lock(&ms_SurfaceQueueLock);
-	_nFree = m_FreeSurfaces.GetCount();
-	_nAlloc = m_AllocatedSurfaces.GetCount();
+	_nFree = (int)m_FreeSurfaces.GetCount();
+	_nAlloc = (int)m_AllocatedSurfaces.GetCount();
 }
 
 void CDX9SubPicAllocator::ClearCache()
